+++ rs/rosetta-api/CHANGELOG.md
@@ -4,7 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.6.0] - Unreleased
+## [1.5.1] - Unreleased
+### Fixed
+- Issue with FOLLOW and NEURON_INFO operations while using hotkeys.
+
+### Changed
+- Controller format (principal or public key) now explicit in operations used with a hotkey.
 
 ## [1.5.0] - 2022-04-06
 ### Added